# dai-front/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DAI — Граф денег API",
    "description": "Роли, кластеры и приоритеты проверки по графу переводов. gid всегда строкой.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/graph": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Get Graph",
        "description": "Весь граф или его срез. Рёбра — только между возвращёнными узлами.",
        "operationId": "get_graph",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Только узлы этого кластера",
              "title": "Cluster Id"
            },
            "description": "Только узлы этого кластера"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "consolidator",
                    "transit",
                    "distributor",
                    "terminal",
                    "coordinator",
                    "peripheral"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Только узлы этой роли",
              "title": "Role"
            },
            "description": "Только узлы этой роли"
          },
          {
            "name": "min_priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1,
              "minimum": 0,
              "description": "Порог priority_score",
              "default": 0.0,
              "title": "Min Priority"
            },
            "description": "Порог priority_score"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "description": "Максимум узлов, по убыванию priority_score",
              "default": 2500,
              "title": "Limit"
            },
            "description": "Максимум узлов, по убыванию priority_score"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Search Nodes",
        "description": "Поиск по началу gid; точное совпадение идёт первым.",
        "operationId": "search_nodes",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Начало строки gid",
              "title": "Q"
            },
            "description": "Начало строки gid"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/nodes/{gid}": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "Get Node",
        "description": "Карточка узла: роль с основанием, входящие и исходящие рёбра, отдельные переводы.",
        "operationId": "get_node",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeCard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/nodes/{gid}/subgraph": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "Get Node Subgraph",
        "description": "Окрестность узла для экрана: узлы в radius шагов (без учёта направления), рёбра — исходные направленные.",
        "operationId": "get_node_subgraph",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gid"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "description": "Сколько шагов от узла в любую сторону",
              "default": 1,
              "title": "Radius"
            },
            "description": "Сколько шагов от узла в любую сторону"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/top": {
      "get": {
        "tags": [
          "priority"
        ],
        "summary": "Get Top Nodes",
        "description": "Ранжированный список приоритетов (top_nodes.csv).",
        "operationId": "get_top_nodes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2248,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/clusters/{cluster_id}": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Get Cluster",
        "description": "Паспорт кластера и его подграф.",
        "operationId": "get_cluster",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClusterDetail": {
        "properties": {
          "cluster": {
            "$ref": "#/components/schemas/ClusterOut"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphResponse"
          }
        },
        "type": "object",
        "required": [
          "cluster",
          "graph"
        ],
        "title": "ClusterDetail"
      },
      "ClusterOut": {
        "properties": {
          "cluster_id": {
            "type": "integer",
            "title": "Cluster Id"
          },
          "n_nodes": {
            "type": "integer",
            "title": "N Nodes"
          },
          "n_seed": {
            "type": "integer",
            "title": "N Seed"
          },
          "sum_kzt_internal": {
            "type": "number",
            "title": "Sum Kzt Internal",
            "description": "Сумма рёбер, у которых оба конца внутри кластера"
          },
          "top_gids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Gids"
          },
          "hypothesis": {
            "type": "string",
            "title": "Hypothesis",
            "description": "Гипотеза о назначении кластера, для проверки"
          }
        },
        "type": "object",
        "required": [
          "cluster_id",
          "n_nodes",
          "n_seed",
          "sum_kzt_internal",
          "top_gids",
          "hypothesis"
        ],
        "title": "ClusterOut"
      },
...
      "EdgeOut": {
        "properties": {
          "src": {
            "type": "string",
            "title": "Src"
          },
          "dst": {
            "type": "string",
            "title": "Dst"
          },
          "sum_kzt": {
            "type": "number",
            "title": "Sum Kzt"
          },
          "n_tx": {
            "type": "integer",
            "title": "N Tx"
          },
          "depth": {
            "type": "integer",
            "title": "Depth",
            "description": "Колено обхода, на котором найдено ребро (1–4)"
          }
        },
        "type": "object",
        "required": [
          "src",
          "dst",
          "sum_kzt",
          "n_tx",
          "depth"
        ],
        "title": "EdgeOut",
        "description": "Направленное ребро: агрегат переводов src → dst за период."
      },
      "GraphMeta": {
        "properties": {
          "mock": {
            "type": "boolean",
            "title": "Mock",
            "description": "True: роли и скоры — заглушка по простым порогам, не результат пайплайна"
          },
          "n_nodes": {
            "type": "integer",
            "title": "N Nodes"
          },
          "n_edges": {
            "type": "integer",
            "title": "N Edges"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "description": "True, если список узлов обрезан параметром limit"
          }
        },
        "type": "object",
        "required": [
          "mock",
          "n_nodes",
          "n_edges",
          "truncated"
        ],
        "title": "GraphMeta"
      },
      "GraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeOut"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeOut"
            },
            "type": "array",
            "title": "Edges",
            "description": "Только рёбра, у которых оба конца есть в nodes"
          },
          "meta": {
            "$ref": "#/components/schemas/GraphMeta"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "meta"
        ],
        "title": "GraphResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeCard": {
        "properties": {
          "node": {
            "$ref": "#/components/schemas/NodeOut"
          },
          "in_edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeOut"
            },
            "type": "array",
            "title": "In Edges"
          },
          "out_edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeOut"
            },
            "type": "array",
            "title": "Out Edges"
          },
          "transfers": {
            "items": {
              "$ref": "#/components/schemas/TransferOut"
            },
            "type": "array",
            "title": "Transfers",
            "description": "Все переводы с участием узла, по дате"
          },
          "pairs": {
            "items": {
              "$ref": "#/components/schemas/TransferPair"
            },
            "type": "array",
            "title": "Pairs",
            "description": "Пары вход→выход узла, по дате входа"
          }
        },
        "type": "object",
        "required": [
          "node",
          "in_edges",
          "out_edges",
          "transfers"
        ],
        "title": "NodeCard",
        "description": "Карточка узла: сам узел, его рёбра, отдельные переводы и пары вход→выход."
      },
      "NodeOut": {
        "properties": {
          "gid": {
            "type": "string",
            "title": "Gid",
            "description": "Идентификатор клиента строкой (int64 в данных)"
          },
          "role": {
            "type": "string",
            "enum": [
              "consolidator",
              "transit",
              "distributor",
              "terminal",
              "coordinator",
              "peripheral"
            ],
            "title": "Role",
            "description": "Основная роль из словаря ТЗ"
          },
          "role_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Role Score",
            "description": "Сила поддержки правила роли, 0–1; не вероятность"
          },
          "cluster_id": {
            "type": "integer",
            "title": "Cluster Id",
            "description": "Номер кластера из clusters.csv"
          },
          "priority_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Priority Score",
            "description": "Приоритет проверки для аналитика, 0–1"
          },
          "evidence": {
            "type": "string",
            "maxLength": 200,
            "title": "Evidence",
            "description": "Почему такая роль: числа из расчёта, до 200 символов"
          },
          "depth": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Depth",
            "description": "Колено обхода, 0 = seed"
          },
          "is_seed": {
            "type": "boolean",
            "title": "Is Seed"
          },
          "in_deg": {
            "type": "integer",
            "title": "In Deg",
            "description": "От скольких разных клиентов получал"
          },
          "out_deg": {
            "type": "integer",
            "title": "Out Deg",
            "description": "Скольким разным клиентам отправлял"
          },
          "in_kzt": {
            "type": "number",
            "title": "In Kzt",
            "description": "Получено внутри графа, KZT"
          },
          "out_kzt": {
            "type": "number",
            "title": "Out Kzt",
            "description": "Отправлено внутри графа, KZT"
          },
          "in_tx": {
            "type": "integer",
            "title": "In Tx"
          },
          "out_tx": {
            "type": "integer",
            "title": "Out Tx"
          },
          "truncated_by_depth": {
            "type": "boolean",
            "title": "Truncated By Depth",
            "description": "Узел на 4-м колене без исходящих: граница выгрузки, не сток"
          },
          "priority_raw": {
            "type": "number",
            "title": "Priority Raw",
            "description": "Сырой скор до нормировки: сумма −ln(доли узлов не ниже) + вес быстрого транзита",
            "default": 0
          },
          "score_terms": {
            "type": "string",
            "title": "Score Terms",
            "description": "Два наибольших вклада в скор, например «in_deg=24 (P99, +7.7)»",
            "default": ""
          },
          "role_checks": {
            "type": "string",
            "title": "Role Checks",
            "description": "Условия назначенной роли с фактическими значениями, до 200 символов",
            "default": ""
          },
          "fast_transit_pairs": {
            "type": "integer",
            "title": "Fast Transit Pairs",
            "description": "Пар вход→выход 1-к-1 с лагом 0–2 дня и отношением сумм 0,8–1,2",
            "default": 0
          },
          "fast_transit_flag": {
            "type": "boolean",
            "title": "Fast Transit Flag",
            "description": "Быстрый транзит: пар не меньше двух",
            "default": false
          },
          "n_seed_upstream": {
            "type": "integer",
            "title": "N Seed Upstream",
            "description": "Из скольких seed узел достижим по исходящим переводам",
            "default": 0
          },
          "betweenness": {
            "type": "number",
            "title": "Betweenness",
            "description": "Посредничество в графе выгрузки",
            "default": 0
          },
          "next_request": {
            "type": "string",
            "title": "Next Request",
            "description": "Какой запрос данных закрыл бы главный пробел по узлу",
            "default": ""
          },
          "limitations": {
            "type": "string",
            "title": "Limitations",
            "description": "Ограничения данных по узлу через «; »; пусто, если их нет",
            "default": ""
          },
          "n_terms_above_p95": {
            "type": "integer",
            "title": "N Terms Above P95",
            "description": "Сколько из 7 признаков скора в верхних 5 % (вклад ≥ −ln 0,05)",
            "default": 0
          },
          "in_queue": {
            "type": "boolean",
            "title": "In Queue",
            "description": "В очереди проверки: не меньше двух признаков в верхних 5 %",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "gid",
          "role",
          "role_score",
          "cluster_id",
          "priority_score",
          "evidence",
          "depth",
          "is_seed",
          "in_deg",
          "out_deg",
          "in_kzt",
          "out_kzt",
          "in_tx",
          "out_tx",
          "truncated_by_depth"
        ],
        "title": "NodeOut",
        "description": "Узел графа с ролью, приоритетом и наблюдаемыми метриками внутри выгрузки."
      },
...
      "SearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NodeOut"
            },
            "type": "array",
            "title": "Items",
            "description": "Совпадения по началу строки gid, по убыванию priority_score"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Сколько всего совпадений (items может быть короче из-за limit)"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SearchResponse"
      },
      "TopNode": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "gid": {
            "type": "string",
            "title": "Gid"
          },
          "role": {
            "type": "string",
            "enum": [
              "consolidator",
              "transit",
              "distributor",
              "terminal",
              "coordinator",
              "peripheral"
            ],
            "title": "Role"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "why": {
            "type": "string",
            "title": "Why",
            "description": "Обоснование текстом"
          },
          "in_queue": {
            "type": "boolean",
            "title": "In Queue",
            "description": "В очереди проверки: не меньше двух признаков в верхних 5 %",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "rank",
          "gid",
          "role",
          "priority_score",
          "why"
        ],
        "title": "TopNode"
      },
      "TopResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TopNode"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TopResponse"
      },
      "TransferOut": {
        "properties": {
          "src": {
            "type": "string",
            "title": "Src"
          },
          "dst": {
            "type": "string",
            "title": "Dst"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "sum_kzt": {
            "type": "number",
            "title": "Sum Kzt"
          }
        },
        "type": "object",
        "required": [
          "src",
          "dst",
          "date",
          "sum_kzt"
        ],
        "title": "TransferOut",
        "description": "Отдельный перевод; дата с точностью до дня, порядок внутри дня неизвестен."
      },
      "TransferPair": {
        "properties": {
          "in_src": {
            "type": "string",
            "title": "In Src",
            "description": "Отправитель входящего перевода (gid строкой)"
          },
          "out_dst": {
            "type": "string",
            "title": "Out Dst",
            "description": "Получатель исходящего перевода (gid строкой)"
          },
          "in_date": {
            "type": "string",
            "format": "date",
            "title": "In Date"
          },
          "out_date": {
            "type": "string",
            "format": "date",
            "title": "Out Date"
          },
          "in_sum": {
            "type": "number",
            "title": "In Sum"
          },
          "out_sum": {
            "type": "number",
            "title": "Out Sum"
          },
          "lag_days": {
            "type": "integer",
            "title": "Lag Days",
            "description": "День выхода − день входа; отрицательный: выход раньше входа"
          },
          "chronology_status": {
            "type": "string",
            "enum": [
              "вход раньше выхода",
              "тот же день, порядок неизвестен",
              "выход раньше входа"
            ],
            "title": "Chronology Status"
          },
          "matched_1to1": {
            "type": "boolean",
            "title": "Matched 1To1",
            "description": "Пара вошла в быстрый транзит (сопоставление 1-к-1, лаг 0–2 дня)"
          }
        },
        "type": "object",
        "required": [
          "in_src",
          "out_dst",
          "in_date",
          "out_date",
          "in_sum",
          "out_sum",
          "lag_days",
          "chronology_status",
          "matched_1to1"
        ],
        "title": "TransferPair",
        "description": "Пара переводов через узел: вход A→узел и выход узел→C, A ≠ C.\n\nОтношение сумм out/in в [0,8; 1,2], |лаг| ≤ 3 дня. Даты с точностью до дня."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}